requirements.txt
numpy>=1.20.0
discord.py>=2.0.0
aiosqlite>=0.17.0
requests>=2.25.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
textual>=0.50.0
